requirements.txt
python-telegram-bot==20.3
telethon==1.29.1
requests
